Moved the docker login to `deploy` stage.
We would need the docker login only when we push the container images to
our public respoitories. This event would happen only during a release
which should never be performed from a forked repository.

Moving the `docker login` to the deploy stage would thus ensure that the
encrypted environment variables are never accessed from a forked repo.