Added part I of the unit-tests for `k8scontext` package.
The Kubernetes go-client provides a package for mocking the Kubernetes API server. We use this mock framework to introduce Ginkgo/Gomega based test framework for the package. This is the first part of a series of commits to introduce unit-tests for this package. In this commit we only test the ability of the controller to listen to CRUD operations to Ingress Resources. We also test that it only reacts to CRUD events on Ingress Resources that have the right annotations set.

Fixed vet.