Added a target to create a helm-releases.
We need to run the `release.sh` script which has dependencies on helm. Have updated the docker dev container with the dependencies.
As a follow up we probably need a script generated that would take a tag and generate the right YAML files and helm release as part of the github tags.