reconcile: add opt-in reconciler event (#775)
This PR provides the ability to reconcile Application Gateway based on a ticker event. Period for the ticker can be modified through helm/environment variable.

Since Application Gateway provides default for settings that were not set by AGIC, this leads to differences in the Generated config vs the config that is returned by ARM. To remove these differences, this PR also provides default values to those properties.