probe: handle hostname extension (#742)
This PR adds code to handle hostname extension annotation for the health probe host.
Probe host is derived from ingress/rule/host if readiness or liveness probe is not provided.

If there are more then 1 host name, then select the one that doesn't have a `*` or `?`. If none is found, then leave host field with the default value.

This is also responsible for intermittent failures in tests.