Give the ability for custom annotations on pod
Many services e.g. datadog use custom annotations on pods to provide
extra functionality. This change to the helm chart allows you to specify
such annotations.

If it is not set you get the default annotations and if it is set you
get the default annotations plus the extras you specified